apps/requirements.txt
numpy==1.24.2
pandas==1.5.3
pyarrow==11.0.0
python-dotenv==0.21.1
